update useControlledDropdownMenu hook to encapsulate logic
